// src/types/typesVM.ts
export enum Coin {
  One = '1₽',
  Two = '2₽',
  Five = '5₽',
  Ten = '10₽'
}

export enum Products {
  Tea = 'Чай',
  Coffee = 'Кофе',
  CoffeeMilk = 'Кофе с молоком',
  Juice = 'Сок'
}

export enum Price {
  Tea = 13,
  Coffee = 18,
  CoffeeMilk = 21,
  Juice = 35
}